fixed files

+++ includes/libft.h
@@ -13,7 +13,7 @@ int ft_atoi(char *str);
 char *ft_strcat(char *dest, const char *src);
 int ft_strcmp(char *s1, char *s2);
 char *ft_strcpy(char *dest, char *src);
-int ft_strlen(char *str);
+size_t ft_strlen(char *str);
 char *ft_strdup(char *src);
 char *ft_strrev(char *str);
 void ft_swap(int *a, int *b);
